requirements.txt
# --- Core AI & LangChain ---
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0  # Crucial for Gemini 2.5

# --- Vector Database ---
faiss-cpu==1.9.0               # The database engine
sentence-transformers==3.1.0   # To create embeddings

# --- PDF & Image Processing (The "Multimodal" part) ---
pymupdf==1.24.10               # Best PDF tool (aka 'fitz')
pillow==10.4.0                 # Image processing library

# --- Utilities ---
python-dotenv==1.0.1           # To read the .env file
requests==2.32.0               # To download papers if needed
numpy==1.26.4